improved test coverage
