+++ src/actions/__tests__/profileAction.spec.js
@@ -7,8 +7,12 @@ import {
   getProfileStart,
   getProfileSuccess,
   editProfile,
+  getArticles,
+  getArticlesStart,
+  getArticlesSuccess,
 } from '../profileAction';
 
+jest.mock('../profileFunc');
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 const mock = new MockAdapter(axios);
@@ -68,5 +72,53 @@ describe('profile action', () => {
       };
       expect(data).toEqual(expectedAction);
     });
+  });
+});
+
+describe('profile action', () => {
+  localStorage.setItem('haven', 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjUsInVzZXJuYW1lIjoibm9ub2p1IiwiZW1haWwiOiJhYkBnbWFpbC5jb20iLCJpYXQiOjE1Njc1MzMwMTcsImV4cCI6MTU2NzYxOTQxN30.YYgbnLLX0hRzxSpSPLR5wj04lzJnTjUKBZlRh_-T9CI');
+  afterEach(() => {
+    mock.reset();
+  });
+  it('success', () => {
+    const expectedAction = {
+      type: 'GET_ARTICLE_SUCCESS',
+      payload: {
+        title: 'article title',
+        description: 'article description',
+      },
+    };
+    expect(getArticlesSuccess(expectedAction.payload)).toEqual(expectedAction);
+  });
+
+  it('success', () => {
+    const expectedAction = {
+      type: 'GET_ARTICLE_START',
+    };
+    expect(getArticlesStart()).toEqual(expectedAction);
+  });
+
+  it('test', () => {
+    const store = mockStore({ });
+    store.dispatch(getArticlesStart);
+    return store.dispatch(getArticles()).then((data) => {
+      const expectedAction = {
+        type: 'GET_ARTICLE_SUCCESS',
+        payload: [],
+      };
+      expect(data).toEqual(expectedAction);
+    });
+  });
+});
+
+describe('Async action creators', () => {
+  it('Should create SIGN_IN_USER_SUCCEEDED when signIn user has been done', async () => {
+    const article = {
+      title: 'image title',
+      description: 'image description',
+    };
+    const dispatch = jest.fn();
+    await getArticles(article)(dispatch);
+    expect(dispatch).toHaveBeenCalledTimes(2);
   });
 });
